requirements.txt
python-telegram-bot[job-queue]==21.0.1
python-dotenv==1.0.1
google-generativeai==0.8.3
psycopg2-binary==2.9.9
Pillow==10.3.0
python-telegram-bot[webhooks]
